% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allR2.R
\name{allR2}
\alias{allR2}
\title{R2 for RDA or CCA Hierarchy}
\usage{
allR2(Y, X, type)
}
\arguments{
\item{Y}{Response variables, typically of community data matrix.}

\item{X}{Explanatory variables, typically of environmental variables.}

\item{type}{The type of constrained ordination: RDA or CCA, the default is "RDA".}
}
\value{
a list containing

\code{gfs} A data frame listing all combinations of explanatory variables in the first column in ascending order, and the corresponding R2 for the model using those variables.

\code{gfsa} A data frame listing all combinations of explanatory variables in the first column in ascending order, and the corresponding adjusted R2 for the model using those variables.
}
\description{
Internal functions for rdacca.hp'. Calculates R2 for RDA or CCA of a single explanatory variable to all combinations of N explanatory variables.
}
\details{
This function calculates R2 and adjusted R2 for the entire hierarchy of RDA or CCA using all combinations of N 
explanatory variables, and returns them as an ordered list ready for input into the function partition.rda. 
This function requires the gtools package in the gregmisc bundle.
}
\examples{
require(ade4)
data(doubs)
spe<-doubs$fish
env<-doubs$env
#Remove empty site 8
spe<-spe[-8,]
env<-env[-8,]
#Hellinger-transform the species dataset for RDA
spe.hel <- decostand(spe, "hellinger")
#select three variables: alt,oxy,and bdo as main explanatory set, via forward selection.
allR2(spe.hel,env[,c("alt","oxy","bdo")],type="RDA")
allR2(spe,env[,c("alt","oxy","bdo")],type="CCA")
}
\references{
Chevan Albert and Sutherland Michael. 1991. Hierarchical Partitioning. The American Statistician.45:90~96

Chris Walsh and Ralph Mac Nally 2013. hier.part: Hierarchical Partitioning. R package version 1.0-4.https://CRAN.R-project.org/package=hier.part
}
\author{
{Jiangshan Lai} \email{lai@ibcas.ac.cn}
}
