% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CanonicalFunctions.r
\name{Canonical.Rsq}
\alias{Canonical.Rsq}
\title{Internal function for rdacca.hp()to caculate the Rsquared,adjusted Rsquared, Pseudo-F value, AIC and AICc for cannocial analysis (RDA,db-RDA and CCA)}
\usage{
Canonical.Rsq(dv, iv, method = "RDA", n.perm = 1000)
}
\description{
Internal function for rdacca.hp()to caculate the Rsquared,adjusted Rsquared, Pseudo-F value, AIC and AICc for cannocial analysis (RDA,db-RDA and CCA)
}
\keyword{internal}
