% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creatbin.r
\name{creatbin}
\alias{creatbin}
\title{Internal function for rdacca.hp() to create diagonal matrix}
\usage{
creatbin(col, binmatrix)
}
\arguments{
\item{col}{Imput number.}

\item{binmatrix}{Imput empty matrix.}
}
\description{
Internal function for rdacca.hp() to create diagonal matrix
}
\keyword{internal}
