% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genList.r
\name{genList}
\alias{genList}
\title{Internal function for rdacca.hp()}
\usage{
genList(ivlist, value)
}
\arguments{
\item{ivlist}{The names of explanatory variables.}

\item{value}{The sequence ID.}
}
\description{
Internal function for rdacca.hp()
}
