% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creatbin.r
\name{odd}
\alias{odd}
\title{Internal function for rdacca.hp() to determine whether the odd number}
\usage{
odd(val)
}
\arguments{
\item{val}{Imput number.}
}
\description{
Internal function for rdacca.hp() to determine whether the odd number
}
