% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition.R
\name{partition.rda}
\alias{partition.rda}
\title{Internal functions for rdacca.hp}
\usage{
partition.rda(gfs, pcan, var.names = NULL)
}
\arguments{
\item{gfs}{a vector of r-squared from a hierarchy of RDA or CCA based on pcan variables in ascending order (as produced by function hpmatrix, but also including the null model as the first element)}

\item{pcan}{the number of variables from which the hierarchy was constructed (maximum =9)}

\item{var.names}{an array of pcan variable names, if required}
}
\description{
Internal functions for rdacca.hp
}
\author{
{Jiangshan Lai} \email{lai@ibcas.ac.cn}
}
