% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permu.hp.r
\name{permu.hp}
\alias{permu.hp}
\title{Permutation Test for Hierarchical Partitioning for Redundancy Analysis and Canonical Correspondence Analysis}
\usage{
permu.hp(Y, X, type = "RDA", permutations = 999)
}
\arguments{
\item{Y}{Community data matrix for RDA and CCA,or a vector for general linear regression model.}

\item{X}{Constraining matrix less than 12 columns, typically of environmental variables.}

\item{type}{the Constrained ordination: RDA or CCA, default "RDA"}

\item{permutations}{the number of permutations required}
}
\value{
a list containing

\item{R2}{unadjusted R-squared for RDA or CCA  for overall model.}

\item{hp.R2}{a dataframe for unadjusted R-squared for individual environmental variables and p-value.}

\item{adj.R2}{adjusted R-squared for RDA or CCA for overall model.}

\item{hp.adjR2}{a dataframe for adjusted R-squared for individual environmental variables and p-value.}
}
\description{
This function performs permutation test for the individual contribution of each environmental variable for Redundancy Analysis and Canonical Correspondence Analysis,
applying the hierarchy algorithm of Chevan and Sutherland (1991) .
}
\details{
This function perform significance test for individual R2 of each independent variable via random permutation of the
rows of the Y data matrix. the test statistic is defined as individual R2. 
At this stage, permu.hp  the defined permutation scheme
}
\examples{
require(ade4)
data(doubs)
spe<-doubs$fish
env<-doubs$env
#Remove empty site 8
spe<-spe[-8,]
env<-env[-8,]
#Remove the 'dfs' variable from the env data frame
env <- env[, -1]
# Hellinger-transform the species dataset
spe.hel <- decostand(spe, "hellinger")
#Forward selection for RDA using vegan's ordiR2step()
mod0 <- rda(spe.hel~ 1, env)  # Model with intercept only
mod1 <- rda(spe.hel~ ., env)
ordiR2step(mod0, mod1,direction = "forward")
#the remaining variables: alt,oxy,bdo
permu.hp(spe.hel,env[,c("alt","oxy","bdo")],type="RDA",permutations=999)
#Forward selection for CCA using vegan's ordiR2step()
mod0 <- cca(spe ~ 1, env)  # Model with intercept only
mod1 <- cca(spe ~., env)
ordiR2step(mod0, mod1,direction = "forward")
#the remaining variables: alt, oxy,bdo
permu.hp(spe,env[,c("alt","oxy","bdo")],type="CCA",permutations=999)
data(mite)
data(mite.env)
mite.hel <- decostand(mite, "hellinger")
mod0 <- rda(mite.hel ~ 1, mite.env)  # Model with intercept only
mod1 <- rda(mite.hel~ ., mite.env)
ordiR2step(mod0, mod1,direction = "forward")
#the remaining variables: WatrCont,Shrub, Substrate,Topo
permu.hp(mite.hel,mite.env[,-1],type="RDA",permutations=999)
#Forward selection using vegan's ordiR2step()
mod0 <- cca(mite ~ 1, mite.env)  # Model with intercept only
mod1 <- cca(mite~., mite.env)
ordiR2step(mod0, mod1,direction = "forward")
#the remaining variables: WatrCont,SubsDens,Substrate,Topo
permu.hp(mite,mite.env[,-4],type="CCA",permutations=99)
}
\references{
Chevan, A. and Sutherland, M. 1991. Hierarchical Partitioning. The American Statistician 45:90~96
Chris Walsh and Ralph Mac Nally 2013. hier.part: Hierarchical Partitioning. R package version 1.0-4.https://CRAN.R-project.org/package=hier.part
}
