% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdar2.R
\name{rdar2}
\alias{rdar2}
\title{An internal function used in rdacca.hp(): Returns only the raw Rsquare and the rank of constraints in RDA, modify "simpleRDA2" from vegan package}
\usage{
rdar2(dv, iv)
}
\arguments{
\item{dv}{Response matrix.}

\item{iv}{Explanatory matrix}
}
\description{
An internal function used in rdacca.hp(): Returns only the raw Rsquare and the rank of constraints in RDA, modify "simpleRDA2" from vegan package
}
\keyword{internal}
