% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpleRDA2.R
\name{simpleRDA2}
\alias{simpleRDA2}
\title{An internal function used in rdacca.hp(): Returns only the raw Rsquare and the rank of constraints in RDA, modify from vegan package}
\usage{
simpleRDA2(Y, X, SS.Y, ...)
}
\arguments{
\item{Y}{Response matrix.}

\item{X}{Explanatory matrix}

\item{SS.Y}{Sum of square of response matrix (Y).}
}
\description{
An internal function used in rdacca.hp(): Returns only the raw Rsquare and the rank of constraints in RDA, modify from vegan package
}
\keyword{internal}
