% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allR2.R
\name{allR2}
\alias{allR2}
\title{Goodness of Fit Measures for RDA or CCA Hierarchy
Calculates goodness of fit measures for RDA or CCA of a single dependent variable to all combinations of N independent variable}
\usage{
allR2(Y, X, type)
}
\arguments{
\item{Y}{Community data matrix.}

\item{X}{Constraining matrix less than 12 columns, typically of environmental variables.}

\item{type}{the Constrained ordination: RDA or CCA, default "RDA"}
}
\value{
\code{gfs} a data frame listing all combinations of independent variables in the first column in ascending order, and the corresponding goodness of fit measure for the model using those variables
}
\description{
Goodness of Fit Measures for RDA or CCA Hierarchy
Calculates goodness of fit measures for RDA or CCA of a single dependent variable to all combinations of N independent variable
}
