% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition.R
\name{partition.rda}
\alias{partition.rda}
\title{Internal functions for rdacca.hp}
\usage{
partition.rda(gfs, pcan, var.names = NULL)
}
\description{
Internal functions for rdacca.hp
}
\author{
{Jiangshan Lai} \email{lai@ibcas.ac.cn}
}
