% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permu.hp.r
\name{permu.hp}
\alias{permu.hp}
\title{Permutation Test for Hierarchical Partitioning for Redundancy Analysis and Canonical Correspondence Analysis}
\usage{
permu.hp(Y, X, type = "RDA", permutations = 999)
}
\arguments{
\item{Y}{Community data matrix.}

\item{X}{Constraining matrix less than 12 columns, typically of environmental variables.}

\item{type}{the Constrained ordination: RDA or CCA, default "RDA"}

\item{permutations}{the number of permutations required}
}
\value{
a list containing

\item{R2}{a dataframe for unadjusted R-squared for individual environmental variables and p-value.}

\item{adjR2}{a dataframe for adjusted R-squared for individual environmental variables and p-value.}
}
\description{
This function performs permutation test for the individual explain percentage of each environmental variable for Redundancy Analysis and Canonical Correspondence Analysis,
applying the hierarchy algorithm of Chevan and Sutherland (1991) .
}
\examples{
require(vegan)
data(varespec)
data(varechem)
rdacca.hp(varespec,varechem[,c("Al","P","K")],pieplot = "tv",type="RDA")
permu.hp(varespec,varechem[,c("Al","P","K")],type="RDA",permutations=999)
rdacca.hp(varespec,varechem[,c("Al","P","K")],pieplot = "tv",type="CCA")
permu.hp(varespec,varechem[,c("Al","P","K")],type="CCA",permutations=999)
}
\references{
Chevan, A. and Sutherland, M. 1991. Hierarchical Partitioning. The American Statistician 45:90~96
}
\author{
{Jiangshan Lai} \email{lai@ibcas.ac.cn}
}
