% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permu.hp.r
\name{permu.hp}
\alias{permu.hp}
\title{Permutation Test for Hierarchical Partitioning for Redundancy Analysis and Canonical Correspondence Analysis}
\usage{
permu.hp(Y, X, type = "RDA", permutations = 999)
}
\arguments{
\item{Y}{Response variables,typically of community data matrix.}

\item{X}{Explanatory variables, typically of environmental variables.}

\item{type}{The type of constrained ordination: RDA or CCA, the default is "RDA".}

\item{permutations}{The number of permutations required.}
}
\value{
a list containing

\item{R2}{Unadjusted R-squared of RDA or CCA for overall model.}

\item{hp.R2}{The independent contribution for each explanatory variable (based on unadjusted  R-squared) and p-value.}

\item{adj.R2}{Adjusted  R-squared of RDA or CCA for overall model.}

\item{hp.adjR2}{The independent contribution for each explanatory variable (based on adjusted  R-squared) and p-value.}
}
\description{
Randomizes row in Y and X and recalculates rdacca.hp permuation times
}
\details{
This function is a permutation routine for the rdacca.hp function which returns a matrix of I values (the independent contribution 
towards explained variation in RDA or CCA) for each explanatory variables. For each permutation, the values in each row (both Y and X) 
are randomized independently, and rdacca.hp is run on the randomized Y and X.  
The function returns a summary table listing the observed I values, the pvalue for each I value.
}
\examples{
require(ade4)
data(doubs)
spe<-doubs$fish
env<-doubs$env
#Remove empty site 8
spe<-spe[-8,]
env<-env[-8,]
#Hellinger-transform the species dataset for RDA
spe.hel <- decostand(spe, "hellinger")
#select three variables: alt,oxy,and bdo as main explanatory set, via forward selection.
permu.hp(spe.hel,env[,c("alt","oxy","bdo")],type="RDA",permutations=99)
permu.hp(spe,env[,c("alt","oxy","bdo")],type="CCA",permutations=99)
}
\references{
Chevan Albert and Sutherland Michael. 1991. Hierarchical Partitioning. The American Statistician.45:90-96

Chris Walsh and Ralph Mac Nally 2013. hier.part: Hierarchical Partitioning. R package version 1.0-4.https://CRAN.R-project.org/package=hier.part
}
\author{
{Jiangshan Lai} \email{lai@ibcas.ac.cn}
}
