% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permu.hp.r
\name{permu.hp}
\alias{permu.hp}
\title{Permutation Test for Hierarchical Partitioning for Redundancy Analysis and Canonical Correspondence Analysis}
\usage{
permu.hp(Y, X, type = "RDA", permutations = 999)
}
\arguments{
\item{Y}{Community data matrix for RDA and CCA,or a vector for general linear regression model.}

\item{X}{Constraining matrix less than 12 columns, typically of environmental variables.}

\item{type}{the Constrained ordination: RDA or CCA, default "RDA"}

\item{permutations}{the number of permutations required}
}
\value{
a list containing

\item{R2}{unadjusted R-squared for RDA or CCA  for overall model.}

\item{hp.R2}{a dataframe for unadjusted R-squared for individual environmental variables and p-value.}

\item{adj.R2}{adjusted R-squared for RDA or CCA for overall model.}

\item{hp.adjR2}{a dataframe for adjusted R-squared for individual environmental variables and p-value.}
}
\description{
This function performs permutation test for the individual contribution of each environmental variable for Redundancy Analysis and Canonical Correspondence Analysis,
applying the hierarchy algorithm of Chevan and Sutherland (1991) .
}
\details{
This function perform significance test for individual R2 of each independent variable via random permutation of the
rows of the Y data matrix. the test statistic is defined as individual R2. 
At this stage, permu.hp  the defined permutation scheme
}
\examples{
require(vegan)
data(varespec)
data(varechem)
rdacca.hp(varespec,varechem[,c("Al","K","N","Baresoil")],pieplot = "tv",type="RDA")
permu.hp(varespec,varechem[,c("Al","K","N","Baresoil")],type="RDA",permutations=999)
rdacca.hp(varespec,varechem[,c("Al","P","K")],pieplot = "tv",type="CCA")
permu.hp(varespec,varechem[,c("Al","P","K")],type="CCA",permutations=999)
}
\references{
Chevan, A. and Sutherland, M. 1991. Hierarchical Partitioning. The American Statistician 45:90~96
Chris Walsh and Ralph Mac Nally 2013. hier.part: Hierarchical Partitioning. R package version 1.0-4.https://CRAN.R-project.org/package=hier.part
}
