% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.rdaccahp.r
\name{plot.rdaccahp}
\alias{plot.rdaccahp}
\title{plot output of rdacca.hp() based on ggplot2}
\usage{
\method{plot}{rdaccahp}(outputList, plot.perc = FALSE)
}
\arguments{
\item{outputList}{the output of rdacca.hp().}

\item{plot.perc}{logical value, if TRUE, the bar plot (based on ggplot2) of the percentage ot independent effects of variables to total Rsquared, the default is FALSE to show plot with original independent effects.}
}
\description{
plot output of rdacca.hp() based on ggplot2
}
\examples{
library(vegan)
data(mite)
data(mite.env)
#Hellinger-transform the species dataset for RDA to deal with the "double zero" problem
mite.hel <- decostand(mite, "hellinger")
rda.hp <- rdacca.hp(mite.hel,mite.env,method="RDA",type="adjR2")
plot(rda.hp)
plot(rda.hp,plot.perc=TRUE)
}
\author{
{Jiangshan Lai} \email{lai@ibcas.ac.cn}

{Pedro Peres-Neto} \email{pedro.peres-neto@concordia.ca}
}
