% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R2current.R
\name{R2current}
\alias{R2current}
\title{Internal functions for rdacca.hp}
\usage{
R2current(Y, current.comb, X, type)
}
\description{
Internal functions for rdacca.hp
}
\author{
{Jiangshan Lai} \email{lai@ibcas.ac.cn}
}
