% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdacca.hp.R
\name{rdacca.hp}
\alias{rdacca.hp}
\title{Hierarchical Partitioning for Redundancy Analysis and Canonical Correspondence Analysis}
\usage{
rdacca.hp(Y, X, type = "RDA", pieplot = "tv")
}
\arguments{
\item{Y}{Community data matrix.}

\item{X}{Constraining matrix less than 12 columns, typically of environmental variables.}

\item{type}{the Constrained ordination: RDA or CCA, default "RDA"}

\item{pieplot}{a pieplot each variable is plotted expressed as percentage of total variation (pieplot="tv") or total explained variation (pieplot="tev").}
}
\value{
a list containing

\item{R2}{unadjusted R-squared for RDA or CCA  for overall model.}

\item{hp.R2}{the individual contribution for each variable (based on unadjusted R-squared).}

\item{adj.R2}{adjusted R-squared for RDA or CCA for overall model.}

\item{hp.adjR2}{the individual contribution for each variable (based on adjusted R-squared).}
}
\description{
This function calculates the individual contribution of each environmental variable for Redundancy Analysis and Canonical Correspondence Analysis,
applying the hierarchy algorithm of Chevan and Sutherland (1991) .
}
\details{
This function calculates the individual contribution of each independent variable to goodness of fit measures on RDA and CCA,
applying the hierarchy algorithm of Chevan and Sutherland (1991) .
all combinations of N independent variable use the function allr2.
It takes the list of goodness of fit measures and, using the partition function,to return a simple table listing each variable, its individuals contribution (I). 
At this stage, the partition routine will not run for more than 12 independent variables. This function requires the vegan,hier.part package.
}
\examples{
require(vegan)
data(varespec)
data(varechem)
#use ordiR2step choose a model by permutation tests in RDA
mod0 <- rda(varespec ~ 1, varechem)  # Model with intercept only
mod1 <- rda(varespec~ ., varechem)
ordiR2step(mod0, mod1)
#the retaining variables: Al,K,N,Baresoil
rdacca.hp(varespec,varechem[,c("Al","K","N","Baresoil")],pieplot = "tv",type="RDA")
rdacca.hp(varespec,varechem[,c("Al","K","N","Baresoil")],pieplot = "tev",type="RDA")
#use ordiR2step choose a model by permutation tests in CCA
mod0 <- cca(varespec ~ 1, varechem)  # Model with intercept only
mod1 <- cca(varespec~ ., varechem)
ordiR2step(mod0, mod1)
#the retaining variables: K,Al,P
rdacca.hp(varespec,varechem[,c("K","Al","P")],pieplot = "tv",type="CCA")
rdacca.hp(varespec,varechem[,c("K","Al","P")],pieplot = "tev",type="CCA")
}
\references{
Chevan, A. and Sutherland, M. 1991. Hierarchical Partitioning. The American Statistician 45:90~96
Chris Walsh and Ralph Mac Nally 2013. hier.part: Hierarchical Partitioning. R package version 1.0-4.https://CRAN.R-project.org/package=hier.part
}
\author{
{Jiangshan Lai} \email{lai@ibcas.ac.cn}
}
