% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdacca.hp.R
\name{rdacca.hp}
\alias{rdacca.hp}
\title{Hierarchical Partitioning for Canonical Correspondence Analysis and Redundancy Analysis}
\usage{
rdacca.hp(Y, X, type = "RDA", pieplot = "tv")
}
\arguments{
\item{Y}{Community data matrix.}

\item{X}{Constraining matrix less than 12 columns, typically of environmental variables.}

\item{type}{the Constrained ordination: RDA or CCA, default "RDA"}

\item{pieplot}{a pieplot each variable is plotted expressed as percentage of total variation (pieplot="tv") or total explained variation (pieplot="tev").}
}
\value{
a list containing

\item{R2}{unadjusted R-squared for RDA or CCA with all environmental variables.}

\item{hp.R2}{the individual contribution for each variable (based on unadjusted R-squared).}

\item{adj.R2}{adjusted R-squared for RDA or CCA with all environmental variables.}

\item{hp.adjR2}{the individual contribution for each variable (based on adjusted R-squared).}
}
\description{
This function calculates the individual explain percentage of each environmental variable for Canonical Correspondence Analysis and Redundancy Analysis,
applying the hierarchy algorithm of Chevan and Sutherland (1991) .
}
\examples{
require(vegan)
data(varespec)
data(varechem)
rdacca.hp(varespec,varechem[,c("Al","P","K")],pieplot = "tv",type="RDA")
rdacca.hp(varespec,varechem[,c("Al","P","K")],pieplot = "tev",type="RDA")
rdacca.hp(varespec,varechem[,c("Al","P","K")],pieplot = "tv",type="CCA")
rdacca.hp(varespec,varechem[,c("Al","P","K")],pieplot = "tev",type="CCA")
}
\references{
Chevan, A. and Sutherland, M. 1991. Hierarchical Partitioning. The American Statistician 45:90~96
}
\author{
{Jiangshan Lai} \email{lai@ibcas.ac.cn}
}
