% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdacca.hp.R
\name{rdacca.hp}
\alias{rdacca.hp}
\title{Hierarchical Partitioning for Redundancy Analysis and Canonical Correspondence Analysis}
\usage{
rdacca.hp(Y, X, pieplot = "tv", type = "RDA", trace = FALSE)
}
\arguments{
\item{Y}{Response variables, typically of community data matrix.}

\item{X}{Explanatory variables, typically of environmental variables.}

\item{pieplot}{A pieplot where each variable is plotted expressed as the percentage of its independent R-squared in the total variation (pieplot="tv") or the total explained variation (pieplot="tev").}

\item{type}{The type of constrained ordination: RDA or CCA, the default is "RDA".}

\item{trace}{If TRUE,R-squared of all combinations (from all.R2 function),joint contribution of each explanatory variable are printed. The default is FALSE.}
}
\value{
a list containing

\item{R2}{Unadjusted R-squared of RDA or CCA for overall model.}

\item{all.R2}{If trace=TRUE,a vector listing the corresponding R-squared for the model using all combinations of each explanatory variables in ascending order.}

\item{hp.R2}{The independent contribution for each explanatory variable (based on unadjusted R-squared).If trace=TRUE, the joint and total R-squared of each explanatory variables is listed.}

\item{adj.R2}{Adjusted R-squared of RDA or CCA for overall model.}

\item{all.adjR2}{If trace=TRUE, a vector listing the corresponding adjusted R-squared for the model using all combinations of each explanatory variables in ascending order.}

\item{hp.adjR2}{The independent contribution for each explanatory variable (based on adjusted R-squared).If trace=TRUE, the joint and total adjusted R-squared of each explanatory variables is listed.}
}
\description{
Partitions variation for each explanatory variable in Redundancy Analysis and Canonical Correspondence Analysis
}
\details{
This function calculates the independent contribution of each explanatory variable to explained variation (R-squared) on RDA and CCA,
applying the hierarchy algorithm of Chevan and Sutherland (1991). The algorithm is that all joint R-squared will be decomposed into equal fractions by number 
of involved explanatory variables and average assigned to these variables. Independent R-squared of each variable will be the sum of assigned R-squared from joint R-squared and unique R-squared. 
All combinations of N explanatory variable use the function allr2().
It takes the list of R-squared and, using the partition function,to return a simple table listing each variable, its independent contribution (I). 
At this stage, the partition routine will not run for more than 9 explanatory variables,due to limitation of computation power and a rounding error for analyses. This code of function is dependent on heir.part packages (Chris and Ralph 2013).
}
\examples{
require(ade4)
data(doubs)
spe<-doubs$fish
env<-doubs$env
#Remove empty site 8
spe<-spe[-8,]
env<-env[-8,]
#Hellinger-transform the species dataset for RDA
spe.hel <- decostand(spe, "hellinger")
#select three variables: alt,oxy,and bdo as main explanatory set, via forward selection.
rdacca.hp(spe.hel,env[,c("alt","oxy","bdo")],pieplot = "tv",type="RDA",trace=TRUE)
rdacca.hp(spe,env[,c("alt","oxy","bdo")],pieplot = "tev",type="CCA")
data(mite)
data(mite.env)
#Hellinger-transform the species dataset for RDA
mite.hel <- decostand(mite, "hellinger")
rdacca.hp(mite.hel,mite.env,pieplot = "tv",type= "RDA",trace=TRUE)
rdacca.hp(mite,mite.env,pieplot = "tev",type= "CCA")
}
\references{
Chevan Albert and Sutherland Michael. 1991. Hierarchical Partitioning. The American Statistician.45:90-96

Chris Walsh and Ralph Mac Nally 2013. hier.part: Hierarchical Partitioning. R package version 1.0-4.https://CRAN.R-project.org/package=hier.part
}
\author{
{Jiangshan Lai} \email{lai@ibcas.ac.cn}
}
